Cleanup temp CF_HOME

+++ src/go/integration/integration_suite_test.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/blang/semver"
@@ -62,7 +63,9 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 
 var _ = SynchronizedAfterSuite(func() {
 	// Run on all nodes
-	Expect(os.RemoveAll(os.Getenv("CF_HOME"))).To(Succeed())
+	if strings.HasPrefix(os.Getenv("CF_HOME"), "/tmp") {
+		Expect(os.RemoveAll(os.Getenv("CF_HOME"))).To(Succeed())
+	}
 }, func() {
 	// Run once
 	Expect(cutlass.RemovePackagedBuildpack(packagedBuildpack)).To(Succeed())
